Refactored other game screens
