+++ src/components/Text.js
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types';
 import useFontSize from "../hooks/useFontSize";
 
 const Text = (props) => {
-    const fontSize = props.style && props.style.fontSize ? useFontSize(props.style.fontSize) : useFontSize(12);;
+    const fontSize = useFontSize(props.style && props.style.fontSize ? props.style.fontSize : 12);
     return (
     <BaseText style={[props.style, {fontSize}]}>
         {props.children}
