+++ src/screens/Game/ExtraPractice.tsx
@@ -2,9 +2,11 @@
 import React from "react";
 import PropTypes from "prop-types";
 import { View, StyleSheet } from "react-native";
+import { NativeStackScreenProps } from "@react-navigation/native-stack";
 import Button from "../../components/Button";
 import Text from "../../components/Text";
 import "react-native-gesture-handler";
+import { RootStackParamList } from "../../types";
 
 const styles = StyleSheet.create({
   root: {
@@ -12,10 +14,10 @@ const styles = StyleSheet.create({
     alignContent: "center",
     justifyContent: "center",
     padding: 20,
-    backgroundColor: 'white'
+    backgroundColor: 'white',
   },
   text: {
-    fontSize: 20,
+    fontSize: 30,
     fontWeight: "bold",
     textAlign: "center",
     marginBottom: 20,
@@ -25,11 +27,13 @@ const styles = StyleSheet.create({
     borderRadius: 10,
   },
   button: {
-    marginTop: 10,
+    marginTop: 15,
   },
 });
 
-function ExtraPractice({ navigation }) {
+type Props = NativeStackScreenProps<RootStackParamList, 'ExtraPractice'>;
+
+function ExtraPractice({ navigation }: Props) {
   return (
     <View style={styles.root}>
       <Text style={styles.text}>Choose an exercise!</Text>
@@ -43,7 +47,7 @@ function ExtraPractice({ navigation }) {
           color: "white",
         }}
         buttonStyle={styles.button}
-        onPress={() => navigation.navigate("MathIntro", { shouldReturn: true })}
+        onPress={() => navigation.navigate("MathIntro", { nextScreenArgs: ["MathMain", { nextScreenArgs: ["ExtraPractice"] }] })}
       />
       <Button
         title="Reading"
@@ -54,9 +58,9 @@ function ExtraPractice({ navigation }) {
           color: "white",
         }}
         buttonStyle={styles.button}
-        onPress={() => navigation.navigate("ReadingIntro", { shouldReturn: true })}
+        onPress={() => navigation.navigate("ReadingIntro", { nextScreenArgs: ["ReadingMain", { nextScreenArgs: ["ExtraPractice"] }] })}
       />
-      <Button
+      {/* <Button
         title="Writing Prompts"
         icon={{
           name: 'pencil',
@@ -65,8 +69,8 @@ function ExtraPractice({ navigation }) {
           color: "white",
         }}
         buttonStyle={styles.button}
-        onPress={() => navigation.navigate("WritingIntro", { shouldReturn: true })}
-      />
+        onPress={() => navigation.navigate("WritingIntro", { nextScreenArgs: ["WritingMain", { nextScreenArgs: ["ExtraPractice"] }] })}
+      /> */}
       <Button
         title="Trivia"
         icon={{
@@ -76,7 +80,7 @@ function ExtraPractice({ navigation }) {
           color: "white",
         }}
         buttonStyle={styles.button}
-        onPress={() => navigation.navigate("TriviaIntro", { shouldReturn: true })}
+        onPress={() => navigation.navigate("TriviaIntro", { nextScreenArgs: ["TriviaMain", { nextScreenArgs: ["ExtraPractice"] }] })}
       />
       <Button
         title="Return to Home"
