--- src/screens/Game/ScoreValues.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const ScoreValues = {
-  correct: 0,
-  total: 0,
-};
-export default ScoreValues;
