+++ src/screens/Stacks/GeneralStacks.tsx
@@ -1,0 +1,16 @@
+import Stack from "./StackNavigator";
+import generalDescriptions from "./generalDescriptions";
+
+const GeneralStacks = generalDescriptions.map(({ name, component, title, options }) => (
+  <Stack.Screen
+    key={name}
+    name={name}
+    component={component}
+    options={{
+      title,
+      ...options,
+    }}
+  />
+));
+
+export default GeneralStacks;
