+++ src/screens/Stacks/MergedStacks.tsx
@@ -1,5 +1,6 @@
 import GameStacks from "./GameStacks";
+import GeneralStacks from "./GeneralStacks";
 
-const MergedStacks = [...GameStacks];
+const MergedStacks = [...GameStacks, ...GeneralStacks];
 
 export default MergedStacks;
